Isolate Method declaration into its own schema
