+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/Closure.scala
@@ -12,12 +12,16 @@ object Closure extends SchemaBase {
       |
       |""".stripMargin
 
-  def apply(builder: SchemaBuilder, base: Base.Schema, enhancements: Enhancements.Schema) =
-    new Schema(builder, base, enhancements)
+  def apply(builder: SchemaBuilder, base: Base.Schema, methodSchema: Method.Schema, enhancements: Enhancements.Schema) =
+    new Schema(builder, base, methodSchema, enhancements)
 
-  class Schema(builder: SchemaBuilder, base: Base.Schema, enhancements: Enhancements.Schema) {
+  class Schema(builder: SchemaBuilder,
+               base: Base.Schema,
+               methodSchema: Method.Schema,
+               enhancements: Enhancements.Schema) {
     import base._
     import enhancements._
+    import methodSchema._
     implicit private val schemaInfo = SchemaInfo.forClass(getClass)
 
     // node properties
